git-svn-id: http://cyberaide.googlecode.com/svn/trunk/project/biostatistics/projects@2595 6884be23-2c2a-0410-b59a-f50d4d03ee1a

+++ fuzzyCMedoidsSeq/fuzzyCMedoids.cpp
@@ -65,7 +65,7 @@ int main(int argc, char** argv) {
 
 		end = clock();
 		//printf("Processing time: %f (ms)\n", ((float)(end - start) / (float)(CLOCKS_PER_SEC)) * (float)1e3);
-		printf("%f\n", (((float)end - (float)start) / (float)(CLOCKS_PER_SEC))) * (float)1e3;
+		printf("%f\n", ((float)(end - start) / (float)(CLOCKS_PER_SEC)) * (float)1e3);
 		oldCost = 1;
 		newCost = 0;
 		//iter = 0;
